Split HexGridView up into BaseGridView and HexGridDrawer.
BaseGridView contains the logic for panning, zooming and moving
pieces, while HexGridDrawer calculates the draw dimensions, draws
the grid and pieces, and translates between grid and screen
coordinates.

The goal is to reuse BaseGridView later to split up RectGridView
similarly, to unify the logic between the two, and add support for
panning and zooming to RectGridView.