Change static AppState instance to be provided by an Application-scoped ViewModelProvider instead.
The advantages are mostly theoretic. First, the AppState will now be initialized lazily, when it's
first requested (by the MainActivity, typically). Second, classes with mutable static state are
often harder to test because they can't be recreated safely (in this case, the last-created
Application instance would overwrite the static field for all instances); this is fairly
inconsequential here since the Application is supposed to be a singleton instance anyway, but it
seems like a good practice to follow on principle.