Hide the action bar by default.
This has the advantage that the puzzle grid can use more space on the screen,
which is valuable in particular for the hex grid puzzle.

The disadvantage is that the user loses some of the options:

  1. Reread the rules.
  2. Skip the tutorial.
  3. Reset puzzle piece positions (without finishing the puzzle first)

I could fix this by adding a floating button or something to show the
context menu (or the app bar itself) but I'm not sure it's all that useful:
the rules are simple and so is the tutorial level, so presumably the player
doesn't need to access them after finishing them once.