Limit zoom level to 2x to mitigate a performance issue.
To reproduce:

  - zoom all the way in
  - drag pieces around
  - zoom all the way out
  - drag pieces around

Dragging is much slower after zooming out than before. I suspect the reason is
that VectorDrawable uses a bitmap cache depending on its dimensions, increasing
the bitmap size when zooming in, but not decreasing it when zooming out.

A better solution might be to draw shapes directly on the canvas without using
bitmaps at all, or resetting the bitmaps when zooming out. But I don't know how
to implement this easily so I'll commit this workaround first.